bot/pyproject.toml
[project]
name = "braindump-bot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3.0",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.0",
]
